requirements.txt
bokeh
numpy
pandas
scipy
matplotlib
pymc3
nodejs
requests==1.1.0
python-firebase
